Refactor UGN-237 change project structure
