+++ src/ReactSpreadsheetImport.tsx
@@ -1,9 +1,9 @@
 import { Modal, ModalContent, ModalOverlay, extendTheme } from "@chakra-ui/react"
-import { ModalCloseButton } from "./ModalCloseButton"
-import { Steps } from "./Steps"
-import { themeOverrides, colorSchemeOverrides } from "../theme"
-import { Providers } from "./Providers"
-import type { RsiProps } from "../types"
+import { ModalCloseButton } from "./components/ModalCloseButton"
+import { Steps } from "./steps/Steps"
+import { themeOverrides, colorSchemeOverrides } from "./theme"
+import { Providers } from "./components/Providers"
+import type { RsiProps } from "./types"
 
 export const theme = extendTheme(colorSchemeOverrides, themeOverrides)
