+++ src/steps/SelectHeaderStep/stories/SelectHeader.stories.tsx
@@ -1,7 +1,7 @@
 import { Box, ChakraProvider, extendTheme } from "@chakra-ui/react"
-import { colorSchemeOverrides, themeOverrides } from "../theme"
-import { headerSelectionTableFields } from "./mockRsiValues"
-import { SelectHeader } from "../components/SelectHeaderStep/SelectHeader"
+import { colorSchemeOverrides, themeOverrides } from "../../../theme"
+import { headerSelectionTableFields } from "../../../stories/mockRsiValues"
+import { SelectHeaderStep } from "../SelectHeaderStep"
 export default {
   title: "Select header",
 }
@@ -11,7 +11,7 @@ const theme = extendTheme(colorSchemeOverrides, themeOverrides)
 export const Table = () => (
   <ChakraProvider theme={theme}>
     <Box display="flex" flexDirection="column" flex={1}>
-      <SelectHeader data={headerSelectionTableFields} onContinue={() => {}} onCancel={() => {}} />
+      <SelectHeaderStep data={headerSelectionTableFields} onContinue={() => {}} onCancel={() => {}} />
     </Box>
   </ChakraProvider>
 )
