+++ src/steps/SelectSheetStep/stories/SelectSheet.stories.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider } from "@chakra-ui/react"
-import { theme } from "../components/ReactSpreadsheetImport"
-import { SelectSheet } from "../components/SelectSheet"
+import { theme } from "../../../ReactSpreadsheetImport"
+import { SelectSheetStep } from "../SelectSheetStep"
 
 export default {
   title: "SelectSheet",
@@ -9,6 +9,6 @@ const sheetNames = ["Sheet1", "Sheet2", "Sheet3"]
 
 export const Basic = () => (
   <ChakraProvider theme={theme}>
-    <SelectSheet sheetNames={sheetNames} onContinue={() => {}} />
+    <SelectSheetStep sheetNames={sheetNames} onContinue={() => {}} />
   </ChakraProvider>
 )
