+++ src/steps/Steps.tsx
@@ -2,7 +2,6 @@ import { UploadFlow } from "./UploadFlow"
 import { Box } from "@chakra-ui/react"
 import { useSteps, Step, Steps as Stepper } from "chakra-ui-steps"
 import { CgCheck } from "react-icons/cg"
-import { dataAttr } from "@chakra-ui/utils"
 
 const steps = [
   { label: "Upload file" },
@@ -14,7 +13,7 @@ const steps = [
 const CheckIcon = ({ color }: { color: string }) => <CgCheck size="2.25rem" color={color} />
 
 export const Steps = () => {
-  const { nextStep, prevStep, setStep, reset, activeStep } = useSteps({
+  const { nextStep, activeStep } = useSteps({
     initialStep: 0,
   })
