+++ src/steps/UploadStep/components/DropZone.tsx
