+++ src/steps/UploadStep/utils/getDropZoneBorder.ts
