--- src/utils/generateOptions.ts
@@ -1,10 +1,0 @@
-export const generateOptions = (tableData: any): any => {
-  if (tableData?.columns) {
-    return tableData!.columns!.map((column: any) => ({
-      ...column,
-      label: column.name || column.uniqueIdentifier,
-    }))
-  } else {
-    return []
-  }
-}
